added some validations

+++ AppAPI/Models/Reservation.cs
@@ -3,11 +3,11 @@
 public class Reservation
 {
     public int ReservationId { get; set; }
-    public long SeatNumber { get; set; }
-    public long UserId { get; set; }
-    public long ProjectionId { get; set; }
+    public int SeatNumber { get; set; }
+    public int UserId { get; set; }
+    public int ProjectionId { get; set; }
 
-    public Reservation(long seatNumber, long userId, long projectionId)
+    public Reservation(int seatNumber, int userId, int projectionId)
     {
         SeatNumber = seatNumber;
         UserId = userId;
